se realizan modificaciones necesarias

+++ src/main/java/com/backend/integrador/dto/domicilio/DomicilioSalidaDTO.java
@@ -1,7 +1,7 @@
 package com.backend.integrador.dto.domicilio;
 
 public class DomicilioSalidaDTO {
-    private int id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
@@ -10,7 +10,7 @@ public class DomicilioSalidaDTO {
     public DomicilioSalidaDTO() {
     }
 
-    public DomicilioSalidaDTO(int id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDTO(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -18,11 +18,11 @@ public class DomicilioSalidaDTO {
         this.provincia = provincia;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
